2012-10-12 15:20 UTC+0700 Andi Jahja <xharbour/AT/telkom/net/id>   * source/vm/fm.c     ! minor formatting
  * source/vm/dlmalloc.c
    ! updated to version 2.8.6, changelog:
        v2.8.6 Wed Aug 29 06:57:58 2012  Doug Lea
        * fix bad comparison in dlposix_memalign
        * don't reuse adjusted asize in sys_alloc
        * add LOCK_AT_FORK -- thanks to Kirill Artamonov for the suggestion
        * reduce compiler warnings -- thanks to all who reported/suggested these

  * contrib/tiff/tif_dirread.c
    ! fixed to compile under MSVC 6

+++ xharbour/contrib/tiff/tif_dirread.c
@@ -3440,7 +3440,8 @@ TIFFReadDirectory(TIFF* tif)
 	if (!dircount)
 	{
 		TIFFErrorExt(tif->tif_clientdata,module,
-		    "Failed to read directory at offset " TIFF_UINT64_FORMAT,(long long unsigned int)nextdiroff);
+		    "Failed to read directory at offset " TIFF_UINT64_FORMAT,(UINT64)nextdiroff);
+		    //"Failed to read directory at offset " TIFF_UINT64_FORMAT,(long long unsigned int)nextdiroff);
 		return 0;
 	}
 	TIFFReadDirectoryCheckOrder(tif,dir,dircount);
@@ -4170,7 +4171,7 @@ TIFFReadCustomDirectory(TIFF* tif, toff_t diroff,
 	if (!dircount)
 	{
 		TIFFErrorExt(tif->tif_clientdata,module,
-		    "Failed to read custom directory at offset " TIFF_UINT64_FORMAT,(long long unsigned int)diroff);
+		    "Failed to read custom directory at offset " TIFF_UINT64_FORMAT,(UINT64)diroff);
 		return 0;
 	}
 	TIFFFreeDirectory(tif);
@@ -4417,14 +4418,14 @@ CheckDirCount(TIFF* tif, TIFFDirEntry* dir, uint32 count)
 		TIFFWarningExt(tif->tif_clientdata, tif->tif_name,
 	"incorrect count for field \"%s\" (" TIFF_UINT64_FORMAT ", expecting %u); tag ignored",
 		    fip ? fip->field_name : "unknown tagname",
-		    (long long unsigned int)dir->tdir_count, count);
+		    (UINT64)dir->tdir_count, count);
 		return (0);
 	} else if ((uint64)count < dir->tdir_count) {
 		const TIFFField* fip = TIFFFieldWithTag(tif, dir->tdir_tag);
 		TIFFWarningExt(tif->tif_clientdata, tif->tif_name,
 	"incorrect count for field \"%s\" (" TIFF_UINT64_FORMAT ", expecting %u); tag trimmed",
 		    fip ? fip->field_name : "unknown tagname",
-		    (long long unsigned int)dir->tdir_count, count);
+		    (UINT64)dir->tdir_count, count);
 		dir->tdir_count = count;
 		return (1);
 	}
